pyproject.toml
[project]
name = "mlops-airlines-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mlflow>=3.0",
    "scikit-learn>=1.5",
    "pandas>=2.2",
    "openai>=1.40",
    "numpy>=1.26",
    "python-dotenv>=1.0",
]
